Backport einiger Entities aus sds
committer: Christopher Jung <bktheg@web.de>

+++ db/tables/user_moneytransfer.sql
@@ -3,7 +3,7 @@ CREATE TABLE `user_moneytransfer` (
   `from` int(11) NOT NULL default '0',
   `to` int(11) NOT NULL default '0',
   `time` int(10) unsigned NOT NULL default '0',
-  `count` int(10) unsigned NOT NULL default '0',
+  `count` bigint(20) unsigned NOT NULL default '0',
   `text` text NOT NULL,
   `fake` tinyint(3) unsigned NOT NULL default '0',
   `type` tinyint(3) unsigned NOT NULL default '0',
